+++ gtsam_unstable/discrete/AllDiff.cpp
@@ -29,7 +29,7 @@ void AllDiff::print(const std::string& s, const KeyFormatter& formatter) const {
 double AllDiff::operator()(const Values& values) const {
   std::set<size_t> taken;  // record values taken by keys
   for (Key dkey : keys_) {
-    size_t value = values.at<size_t>(dkey);  // get the value for that key
+    uint64_t value = values.at<uint64_t>(dkey);  // get the value for that key
     if (taken.count(value)) return 0.0;  // check if value alreday taken
     taken.insert(value);  // if not, record it as taken and keep checking
   }
